Fix a warning on double to int conversion for Motif version. ASSIGNED - bug 1401: Use vertex arrays with OpenGL https://tracker.physiomeproject.org/show_bug.cgi?id=1401
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@6059 3c1deb5b-d424-0410-962d-aba41a686d42

+++ source/emoter/emoter_dialog.cpp
@@ -4223,7 +4223,7 @@ DESCRIPTION :
 				(Movie_graphics_set_Graphics_window(emoter_dialog->movie,
 					graphics_window)))
 			{
-				number_of_frames = floor( emoter_dialog->time_maximum -
+				number_of_frames = (int)floor( emoter_dialog->time_maximum -
 					emoter_dialog->time_minimum + 1);
 				for ( frame = 0 ; frame < number_of_frames ; frame++ )
 				{
